+++ src/modules/scenarioeditor/GridEditView.vue
@@ -452,12 +452,15 @@ function doDelete(e: KeyboardEvent) {
 async function onUnitEdit(unit: NUnit, b: ColumnField, c: string) {
   if (b === "sidc") {
     const newSidcValue = await getModalSidc(c, {
-      symbolOptions: unitActions.getCombinedSymbolOptions(unit),
+      symbolOptions: unit.symbolOptions,
+      inheritedSymbolOptions: unitActions.getCombinedSymbolOptions(unit, true),
+      reinforcedStatus: unit.reinforcedStatus,
     });
     if (newSidcValue !== undefined) {
       updateUnit(unit.id, {
         sidc: newSidcValue.sidc,
         symbolOptions: newSidcValue.symbolOptions,
+        reinforcedStatus: newSidcValue.reinforcedStatus,
       });
     }
   }